Fixed issue with PHP 7 session_regenerate_id()
PHP 7 closes the session where php 5 did not. When a session is closed the database object gets set to null which errors when PHP attempts to open/read a new session with a new ID. Now the database object will never be set to null and will retain its connection.